feat: refactoring fixtures and creating a base fixture reducing duplicated code
