+++ FC.Codeflix.Catalog.UnitTests/Application/Category/Common/CategoryUseCasesBaseFixture.cs
@@ -1,18 +1,25 @@
-﻿using FC.Codeflix.Catalog.Domain.Entity;
+﻿using FC.Codeflix.Catalog.Application.Interfaces;
+using FC.Codeflix.Catalog.Application.UseCases.Category.CreateCategory;
 using FC.Codeflix.Catalog.Domain.Repository;
 using FC.Codeflix.Catalog.UnitTests.Common;
 using Moq;
-using Xunit;
+using UseCase = FC.Codeflix.Catalog.Domain.Entity;
 
-namespace FC.Codeflix.Catalog.UnitTests.Application.GetCategory;
-
-[CollectionDefinition(nameof(GetCategoryTestFixture))]
-public class GetCategoryTestFixtureCollection : ICollectionFixture<GetCategoryTestFixture> { }
-
-public class GetCategoryTestFixture : BaseFixture
+namespace FC.Codeflix.Catalog.UnitTests.Application.Category.Common;
+public abstract class CategoryUseCasesBaseFixture : BaseFixture
 {
+
     public Mock<ICategoryRepository> GetRepositoryMock()
         => new();
+    public Mock<IUnitOfWork> GetUnitOfWorkMock()
+        => new();
+
+    public CreateCategoryInput GetInput()
+    => new(
+        GetValidCategoryName(),
+        GetValidCategoryDescription(),
+        GetRandomBoolean()
+        );
 
     public string GetValidCategoryName()
     {
@@ -34,9 +41,13 @@ public class GetCategoryTestFixture : BaseFixture
         return categoryDescription;
     }
 
-    public Category GetValidCategory()
+    public static bool GetRandomBoolean()
+        => new Random().NextDouble() < 0.5;
+
+    public UseCase.Category GetExampleCategory()
     => new(
         GetValidCategoryName(),
-        GetValidCategoryDescription()
+        GetValidCategoryDescription(),
+        GetRandomBoolean()
     );
 }
