+++ FC.Codeflix.Catalog.UnitTests/Application/Category/DeleteCategory/DeleteCategoryTest.cs
@@ -4,7 +4,7 @@ using Moq;
 using Xunit;
 using UseCase = FC.Codeflix.Catalog.Application.UseCases.Category.DeleteCategory;
 
-namespace FC.Codeflix.Catalog.UnitTests.Application.DeleteCategory;
+namespace FC.Codeflix.Catalog.UnitTests.Application.Category.DeleteCategory;
 
 [Collection(nameof(DeleteCategoryTestFixture))]
 public class DeleteCategoryTest
@@ -23,7 +23,7 @@ public class DeleteCategoryTest
         //arrange
         var repositoryMock = _fixture.GetRepositoryMock();
         var unitOfWorkMock = _fixture.GetUnitOfWorkMock();
-        var categoryExample = _fixture.GetValidCategory();
+        var categoryExample = _fixture.GetExampleCategory();
         repositoryMock.Setup(x => x.Get(categoryExample.Id,
             It.IsAny<CancellationToken>())
         ).ReturnsAsync(categoryExample);
