+++ FC.Codeflix.Catalog.UnitTests/Application/Category/DeleteCategory/DeleteCategoryTestFixture.cs
@@ -1,0 +1,11 @@
+﻿using FC.Codeflix.Catalog.UnitTests.Application.Category.Common;
+using Xunit;
+
+namespace FC.Codeflix.Catalog.UnitTests.Application.Category.DeleteCategory;
+
+[CollectionDefinition(nameof(DeleteCategoryTestFixture))]
+public class DeleteCategoryTestFixtureCollection : ICollectionFixture<DeleteCategoryTestFixture> { }
+
+public class DeleteCategoryTestFixture : CategoryUseCasesBaseFixture
+{
+}
