+++ FC.Codeflix.Catalog.UnitTests/Application/Category/GetCategory/GetCategoryInputValidatorTest.cs
@@ -2,7 +2,7 @@
 using FluentAssertions;
 using Xunit;
 
-namespace FC.Codeflix.Catalog.UnitTests.Application.GetCategory;
+namespace FC.Codeflix.Catalog.UnitTests.Application.Category.GetCategory;
 
 [Collection(nameof(GetCategoryTestFixture))]
 public class GetCategoryInputValidatorTest
