+++ FC.Codeflix.Catalog.UnitTests/Application/Category/UpdateCategory/UpdateCategoryInputValidatorTest.cs
@@ -2,7 +2,7 @@
 using FluentAssertions;
 using Xunit;
 
-namespace FC.Codeflix.Catalog.UnitTests.Application.UpdateCategory;
+namespace FC.Codeflix.Catalog.UnitTests.Application.Category.UpdateCategory;
 
 [Collection(nameof(UpdateCategoryTestFixture))]
 public class UpdateCategoryInputValidatorTest
