+++ FC.Codeflix.Catalog.UnitTests/Application/Category/UpdateCategory/UpdateCategoryTestDataGenerator.cs
@@ -1,11 +1,25 @@
-﻿namespace FC.Codeflix.Catalog.UnitTests.Application.CreateCategory;
-public class CreateCategoryTestDataGenerator
+﻿namespace FC.Codeflix.Catalog.UnitTests.Application.Category.UpdateCategory;
+public class UpdateCategoryTestDataGenerator
 {
+    public static IEnumerable<object[]> GetCategoriesToUpdate(int times = 10)
+    {
+        var fixture = new UpdateCategoryTestFixture();
+        for (int indice = 0; indice < times; indice++)
+        {
+            var exampleCategory = fixture.GetExampleCategory();
+            var exampleInput = fixture.GetValidInput(exampleCategory.Id); ;
+            yield return new object[]
+            {
+                exampleCategory, exampleInput
+            };
+        };
+    }
+
     public static IEnumerable<object[]> GetInvalidInputs(int times = 12)
     {
-        var fixture = new CreateCategoryTestFixture();
+        var fixture = new UpdateCategoryTestFixture();
         var invalidInputsList = new List<object[]>();
-        var totalInvalidCases = 4;
+        var totalInvalidCases = 3;
 
         for (int index = 0; index < times; index++)
         {
@@ -24,12 +38,6 @@ public class CreateCategoryTestDataGenerator
                     });
                     break;
                 case 2:
-                    invalidInputsList.Add(new object[] {
-                        fixture.GetInvalidInputDescriptionNull(),
-                        "Description should not be null"
-                    });
-                    break;
-                case 3:
                     invalidInputsList.Add(new object[] {
                         fixture.GetInvalidInputDescriptionTooLong(),
                         "Description should be less or equal to 10000 characters long"
