--- FC.Codeflix.Catalog.UnitTests/Application/CreateCategory/CreateCategoryTestFixture.cs
@@ -1,78 +1,0 @@
-﻿using FC.Codeflix.Catalog.Application.Interfaces;
-using FC.Codeflix.Catalog.Application.UseCases.Category.CreateCategory;
-using FC.Codeflix.Catalog.Domain.Repository;
-using FC.Codeflix.Catalog.UnitTests.Common;
-using Moq;
-using Xunit;
-namespace FC.Codeflix.Catalog.UnitTests.Application.CreateCategory;
-
-[CollectionDefinition(nameof(CreateCategoryTestFixture))]
-public class CreateCategoryTestFixtureCollection : ICollectionFixture<CreateCategoryTestFixture> { }
-
-public class CreateCategoryTestFixture : BaseFixture
-{
-    public string GetValidCategoryName()
-    {
-        var categoryName = "";
-        while (categoryName.Length < 3)
-            categoryName = Faker.Commerce.Categories(1)[0];
-        if (categoryName.Length > 255)
-            categoryName = categoryName[..255];
-        return categoryName;
-    }
-
-    public string GetValidCategoryDescription()
-    {
-        var categoryDescription =
-            Faker.Commerce.ProductDescription();
-        if (categoryDescription.Length > 10000)
-            categoryDescription =
-                categoryDescription[..10000];
-        return categoryDescription;
-    }
-
-    public static bool GetRandomBoolean()
-        => (new Random().NextDouble() < 0.5);
-
-    public CreateCategoryInput GetInput()
-        => new(
-            GetValidCategoryName(),
-            GetValidCategoryDescription(),
-            GetRandomBoolean()
-            );
-
-    public CreateCategoryInput GetInvalidInputShortName()
-    {
-        var invalidInputName = GetInput();
-        invalidInputName.Name = invalidInputName.Name[..2];
-        return invalidInputName;
-    }
-
-    public CreateCategoryInput GetInvalidInputTooLongName()
-    {
-        var invalidInputTooLongName = GetInput();
-        var invalidName = Faker.Lorem.Letter(256);
-        invalidInputTooLongName.Name = invalidName;
-        return invalidInputTooLongName;
-    }
-
-    public CreateCategoryInput GetInvalidInputDescriptionNull()
-    {
-        var invalidInputDescriptionNull = GetInput();
-        invalidInputDescriptionNull.Description = null!;
-        return invalidInputDescriptionNull;
-    }
-
-    public CreateCategoryInput GetInvalidInputDescriptionTooLong()
-    {
-        var invalidInputDescriptionTooLong = GetInput();
-        var invalidDescription = Faker.Lorem.Letter(11000);
-        invalidInputDescriptionTooLong.Description = invalidDescription;
-        return invalidInputDescriptionTooLong;
-    }
-
-    public Mock<ICategoryRepository> GetRepositoryMock()
-        => new();
-    public Mock<IUnitOfWork> GetUnitOfWorkMock()
-        => new();
-}
